Update event window. Update flink config.
Update flink config to allow it to use more memory.

Limit event window as the patterns accumulated rather quickly after
a while which used up all the available memory. Checking just one or a
few seconds is not intresting for this usecase.

This can probably be handled but seems like it will require a lot of extra
work according to examiner. Will focus on different approaches.